wip: PoC
It mostly works, but needs to be turned into a CLI. Notably, it'd be
nice to be able to change the ports (for both the real and go proxy).
Another nicety would be to be able to change the timeout. We could
also allow the user to provide the real proxy's PID directly, since
right now the port is only used to find the PID.

Ultimately, I'd like to also be able to set this to forward to a
remote proxy, not just a local one. But that /should/ be a lot
simpler than the local case, since we don't have to fuck around
with mapping ports to PIDs and stuff.

And when I'll have infinite money and time, I'll try to integrate
NTLM integration directly, so that future students like me don't
have to use (and, most importantly, install) Cntlm as a middle-man.

Ok that's a lie, I'd rather rewrite it in C first, so that you
don't have to first setup the proxy for the Go toolchain before
being free the shackles of the proxy config clusterfuck.

Oh, also SOCKS5 would be nice.

Anyway, it works right now. Yay!
I'll try and write up a little gist for it :)