feat(reroute): detect proxy pid no matter when it was/is started
our main means of detecting the proxy pid is now to listen for bind()s
and register the process that binds to REAL_PROXY_PORT as needing to
be ignored.

we still need the whole pid_finder thing in go at startup in case
the proxy was started before us, b/c obviously we can't attach to
calls in the past.

this *does* allow us to add bypasses based on src port, so yay!

i've left the old code commented for now, in case it malfunctions,
but thus far it's been running pretty smoothly.